removed inheritance
