--- Assets/Scripts/ShaderContolers/ShadeContoler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fe2f352c5b8771d48aad552d693ff361
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
